Insert CreativeWork schema
Using `Article` as a base class for most things was fine until we
introduced `SearchResultPage`.  `CreativeWork` is the common ancestor
of both of these, so makes sense to introduce.

We did not break anything by having the `SearchResultPage` class inherit
from `Article` because we didn't use anything in `Article` that was not in
`CreativeWork` apart from `articleBody` and that defaulted to empty when our
content items didn't include relevant content. It was a bit of a violation
of OO principles though.

By adding `CreativeWork` here, we hopefully ensure that things won't
break should we subsequently use specialised properties of `Article`.

Note that we are just reflecting the ancestry of https://schema.org/Article
and https://schema.org/SearchResultsPage here and not changing any other
behaviour.
