+++ lib/govuk_publishing_components/presenters/machine_readable/search_results_page_schema.rb
@@ -7,7 +7,7 @@ module GovukPublishingComponents
 
       def structured_data
         # http://schema.org/SearchResultsPage
-        data = ArticleSchema.new(@page).structured_data
+        data = CreativeWorkSchema.new(@page).structured_data
         data["@type"] = "SearchResultsPage"
         data
       end
